requirements.txt
PYYaml==3.13
lxml==4.3.3
bs4